src/lambda/dashboard_api/requirements.txt
boto3>=1.26.0
psycopg2-binary>=2.9.5
